fix: annotate max_age_hours as float in cleanup_old_jobs (#3198)
The parameter was typed as int but received float values (e.g. 1.0
from completed_job_ttl_hours). Fixed both the engine method and the
router endpoint.

Closes #3198

+++ auralis-web/backend/core/processing_engine.py
@@ -633,7 +633,7 @@ class ProcessingEngine:
 
         return False
 
-    async def cleanup_old_jobs(self, max_age_hours: int = 24) -> int:
+    async def cleanup_old_jobs(self, max_age_hours: float = 24) -> int:
         """Clean up old completed jobs and their files.
 
         Protected by _jobs_lock so that concurrent invocations (worker finally-
